// src/app/interfaces/pokemonType.ts
export interface PokemonData {
    name: string;
    url: string;
  };